+++ NEWS.md
@@ -44,6 +44,7 @@ for quite some time.
 + Added more teas.
 + Teas can now be cooled to make iced tea.
 + Fixed octocopter parts appearing more than once.
++ Tea can now be chilled in an icebox to make iced tea.
 
 
 ### Update 2014-6-23 (alpha-0.0.7 release)